+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # @connect/connect-design-system
 
+## 1.18.4
+
+### Patch Changes
+
+- 03d51d0: - New Component: N/A
+  - Updated Component:
+    - `<Icon />`: collapes-left-outline, commonError, expand-right-outline, ml, syl, and udl Icons added
+    - `<Paper />`: Outline prop added
+  - Bug Fix: N/A
+  - Other: Colors unified with new name, Stories updated with new color names.
+
+## 1.18.3
+
+### Patch Changes
+
+- 81d3288: Added type to useState in Icon component.
+
 ## 1.18.2
 
 ### Patch Changes